requirements.txt
anthropic>=0.25.0
langgraph>=0.1.0
langchain-core>=0.2.0
pymupdf>=1.24.0
playwright>=1.44.0
httpx>=0.27.0
streamlit>=1.35.0
pydantic>=2.0.0
python-dotenv>=1.0.0
firecrawl-py>=0.0.16